cmd/govim: first cut of command-line mode to query "parent"
This enables govim to be used with fzf and other such Vim plugins that
require a client program in order to run a command. In the case of fzf,
the client instance of govim is used in order to re-query the parent
instance, which in turn calls on to gopls' workspace symbol search. This
allows the search results within fzf to be refined on each key stroke.

This approach, however, is not specific to fzf. Hence some configuration
of Vim/fzf is required. For a sample .vimrc configuration see:

https://github.com/govim/govim/wiki/vimrc-tips#-tip-use-fzf-with-workspace-symbol-search

Fixes #795